requirements.txt
flask>=3.0.0
python-docx>=1.1.0
pdfplumber>=0.11.0
nltk>=3.8.0
docx2pdf>=0.1.8
